Refactor integer to decimal in tables cards/watches

+++ data/migrations/20201019160715_cards.js
@@ -7,7 +7,7 @@ exports.up = function(knex) {
         .notNullable()
       
       cards
-        .integer('price')
+        .decimal('price')
         .notNullable()
 
       cards
